requirements.txt
feedparser==6.0.11
requests==2.32.3
anthropic==0.42.0
python-telegram-bot==21.9
python-dotenv==1.0.1
